[schools] Move school functions from randomActivityGen.py into school.py
